imageoperations/deleteimage fixed, added vehicle and client comntroller and moved methods from home controller

+++ OtoTamir.WEBUI/Services/ImageOperations.cs
@@ -24,7 +24,11 @@
 
         public static void DeleteImage(string fileName)
         {
-            var path = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\images", fileName);
+            if (fileName == "avatar.png")
+            {
+                return;
+            }
+            var path = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot","images", fileName);
             if (System.IO.File.Exists(path))
             {
                 System.IO.File.Delete(path);
